Add thread safety extension to OpenSHMEM
This patch adds basic thread safety that has been proposed to OpenSHMEM.
This must be enabled at configure time and is enabled in the application
by calling the new shmem_init() function and requesting a level of
thread safety higher than SHMEM_THREAD_SINGLE.

This is a first pass at implementing this functionality, focused on
correctness.  A future commit will replace some of the mutexes that are
used with atomic operations.