Convert free list to use mutex interface
This converts the shmem_free_list functions over to the SHMEM_MUTEX_...
functions that are configured based on the level of threading that the
user has enabled.  Previous to this patch, the free lists directly used
spinlocks all of the time.  This should yield a performance increase for
non-threaded executions.