Update travis to build OFI and run tests
Update Travis CI script to build OFI and its dependencies.  The unit
tests and UH test suite are also run against the OFI build.

Unfortunately, Portals on sockets is buggy; there is code in place to
run those tests, but it's disabled because it generates many failures
and sometimes causes the testing process to hang.

Signed-off-by: James Dinan <james.dinan@intel.com>