global_exit is busted
Global exit succeeds at tearing down the execution, but it's sloppy and
does not sufficiently absorb teardown errors from the transport layer.
As a result, the test works (i.e., it exits), but it does not output a
success error code like it should.

Moving this test to XFAIL temporarily.

Signed-off-by: James Dinan <james.dinan@intel.com>