Fix race in shmem_lock_clear
Correct a race in clear_lock that occurred in the sequence (1) check if
next PE pointer (lock->data) was updated (2) else wait for update.  In
the second step, lock->data was re-read rather than using the value from
the first step.  As a result, if an update arrived between steps,
shmem_lock_clear could deadlock.

Signed-off-by: James Dinan <james.dinan@intel.com>