Add support for long put frags and reduce latency
Previous to this patch, all puts above the inject limit generated a frag
object with one frag.  This patch adds a path for messages above the
inject threshold and below the max message size with a counting event
rather than fragmentation and full event.  This should improve put
latency noticably.

Signed-off-by: James Dinan <james.dinan@intel.com>