Update recdbl psync atomic to use SUM/INT
The recursive doubling collectives used atomic sum on signed byte in
their psync updates.  This added a dependence on the transport layer
supporting atomic sum on byte, which caused a portability issue for the
GNI provider.

This update replaces signed byte with signed int, which results in a
larger psync; however, it still fits in a cache line on systems with
32-bit integers.

Signed-off-by: James Dinan <james.dinan@intel.com>