add support for large page sheap for linux
With this commit, support is added to to SOS
for allowing the symmetric heap to be backed by
large pages on Linux.

This feature is not reliant on THP feature in Linux.

The use of default large page size (2MB) is triggered
by setting the

SMA_SYMMETRIC_HEAP_USE_HUGE_PAGES

environment variable.

For systems that support multiple large pages (such
as the synthetic large pages available on Cray systems),
there is an additional env. variable

SMA_SYMMETRIC_HEAP_PAGE_SIZE

It is aware of MB, GB, etc. syntax.  So, to use the
128MB large pages on Cray system the env. variable
would be set as

export SMA_SYMMETRIC_HEAP_PAGE_SIZE=128MB

Signed-off-by: Howard Pritchard <howardp@lanl.gov>