Simplify transport configury behavior
With this change, users may pass either 1) --with-ofi, 2)
--with-portals4, 3) neither, and 4) *not* both.  If both, raise an error.
If neither, set transport to "none" and warn the user.  Similar for
XPMEM and CMA; but if neither, disable XPMEM and CMA and print result
(with no warning).  This change prevents automatic detection of these
libraries, but prevents strange linking issues with certain system-wide
installations.

Signed-off-by: David Ozog <david.m.ozog@intel.com>