Suppress "unused" warnings in c11 bindings generator
These variables are "used" statically to steer the _Generic type
selection expression.  Some compilers (e.g. gcc 6, but not 7) may warn
that the variable is not used (since it is not referenced in the dynamic
execution of the program).  This patch applies the unused attribute to
these variables to suppress such warnings.

Signed-off-by: James Dinan <james.dinan@intel.com>