Quiet CMA build warnings
Move _GNU_SOURCE declaration to config.h to quiet build warnings.  This has
broad consqeuences, since the _GNU_SOURCE feature macro is now defined across
the whole build, but is unfortunately necessary to use the libc CMA interface
since the first #include of a system header may pull in the header containing
process_vm_readv/writev.  If this breaks for some reason, the fix is probably
to use the syscall interface instead of the libc bindings.

Signed-off-by: James Dinan <james.dinan@intel.com>