CMA: Move _GNU_SOURCE definition to top-level
Because the CMA transport header depends on _GNU_SOURCE, it won't work
to define it in the header itself (since _GNU_SOURCE needs to be defined
before any system headers are included and transport_cma.h is not
reliably the first header included).  So, move the definiton of
_GNU_SOURCE to config.h and flag an error if it's not defined when
transport_cma.h is included.

Signed-off-by: James Dinan <james.dinan@intel.com>