Portals: Fix LE counter in CMA/XPMEM builds
Starting with SOS v1.3.0, the LE counter was not supplied when hard
polling was enabled (since CT events are not enabled on the LE).  This
causes the Portals 4 reference implementation to crash (even when the CT
handle is given as PTL_HANDLE_INVALID rather than undefined).  This
patch puts back the target counter for hard polling builds and captures
the counter disabling code using a different preprocessor gate, in case
we find a use for this in the future.

Signed-off-by: James Dinan <james.dinan@intel.com>