CI: Ignore failures during UCX testing
There two issues with UCX in CI #1148 and #1048 that are under debug.
Temporarily ignore the failures in UCX until these tickets are root caused and
resolved inorder to reduce noise in CI results.

All of the UCX tests are run, but erroneous results for UCX's "make check" are
converted into warnings and are annotated in the results output.

Signed-off-by: Mark F. Brown <mark.f.brown@intel.com>